+++ src/routes/home/Home.js
@@ -11,7 +11,10 @@ import React, { PropTypes } from 'react';
 import withStyles from 'isomorphic-style-loader/lib/withStyles';
 import s from './Home.scss';
 
-function Home({ news }) {
+const title = 'React Starter Kit';
+
+function Home({ news }, context) {
+  context.setTitle(title);
   return (
     <div className={s.root}>
       <div className={s.container}>
@@ -39,5 +42,6 @@ Home.propTypes = {
     contentSnippet: PropTypes.string,
   })).isRequired,
 };
+Home.contextTypes = { setTitle: PropTypes.func.isRequired };
 
 export default withStyles(Home, s);