+++ src/routes/register/index.js
@@ -10,9 +10,12 @@
 import React from 'react';
 import Register from './Register';
 
-export const path = '/register';
-export const action = async (state) => {
-  const title = 'New User Registration';
-  state.context.onSetTitle(title);
-  return <Register title={title} />;
+export default {
+
+  path: '/register',
+
+  action() {
+    return <Register />;
+  },
+
 };